Clean up full_check a bit and move the file
